Account for fixed header in in-page anchor links
This solution is kind of hacky (basically everything below the header
becomes its own pane, with its own scrollbar) but it's imo better than
the alternative of adding a top padding and negative top margin for
anything that could be an anchor, so alas.

The ?1 query param for the stylesheet is just to force a cache refresh.